refactor(icons): move ComponentIcon to icons.tsx for better modularity
feat(command-menu): add dynamic icon rendering for components

chore(page.tsx): comment out unused ComponentIcon code to clean up file
